requirements.txt
asyncio-mqtt @ git+https://github.com/wirenboard/python-asyncio-mqtt@dev/v0.10.0
json-rpc==1.13.0
jsonschema==3.2.0
mqttrpc @ git+https://github.com/wirenboard/python-mqtt-rpc@main
paho-mqtt==1.5.1
python-dali @ git+https://github.com/wirenboard/python-dali@dev/v0.11
websockets==8.1
